Distance Code w/ accurate light detection
Run ObjRecognitionController class (might need to use eclipse, working on a solution). Requires OpenCV 4.2.0 32 bit and OpenCV 3.4.9 32 bit as the dependencies, as well as JavaFX library (for GUI output). Set hue range to 0-180, saturation to 240-255, and value to 240-255 to match bright light in the GUI sliders (or edit the ObjRecognition.fxml file to get ranges).